add public reviews and average ratings feature
at home page:
- add the average rating of an inn. if an
  inn doesn't have any reviews, the field
  will be left blank

at inn details page
- add a section with the three latest
  reviews' comments
- display the average score of an inn
- display the amount of reviews an inn
  has received all time
- display a link to the public reviews
  list page

add public reviews list page
- display the review author's first name
- display the review's comment
- also display the reviews count and
  average score

as a development decision, the host will
not manage its reviews through the public
reviews list, leaving the my_inn_reviews
routes as it is. this facilitates to use
different communications with regular users
and the host, and keep the sensible actions
out of reach of malicious users.