add rooms list endpoint to api
add a endpoint to return a list of rooms
when receiving a request with a existing
inn_id parameter.

only returns active rooms, raises an error
if the inn doesn't exist and returns an
empty array if the inn does not have any
active rooms.

api documentation at README will be updated
in later commits.