add inn details endpoint to api
add an endpoint to return the details
of an inn when receiving a request with
a existing inn_id parameter.

can return inactive inns, raises an error
if inn doesn't exists.

it also returns the average rating of the
inn if it has reviews. if contrary, returns
an empty string.

needed to add a few methods in the api
controller to make it work, i'll ask in
the next call if they can be moved to
the model.