add check room availability endpoint to api
create endpoint to check if a room is available
within the paramters provided on request. using
the application logic, it tries to create a
pending reservation and if succeeds, indicates
that the room is available.

- if the room is available, the pending reservation
  data is provided in the response as a JSON file,
  with the price of the selected period
- if the room is unavailable, return an hash with
  the full errors message as the value, informing
  the user why the attempt failed
- returns 404 if the provided room_id does not
  exist
- returns 500 if an internal server error occurs
  during the request

also in this commit the jbundler gem is turned
off due to its impacts on the testing process.