add consumables features for hosts
hosts can register consumables in
a reservation to expand their revenue
options as an inn owner.

description and price must be provided
for a consumable to be valid.

an invoice model will be created in
following commits to calculate the total
price and enable a better separation of
the logics.